BB-1598: There is no way to delete attachment if there are more than one Doctrine flush  - fixed issue  - updated unit test

+++ src/Oro/Bundle/AttachmentBundle/Manager/AttachmentManager.php
@@ -109,7 +109,7 @@ class AttachmentManager
     public function preUpload(File $entity)
     {
         if ($entity->isEmptyFile()) {
-            if ($this->filesystem->has($entity->getFilename())) {
+            if ($entity->getFilename() && $this->filesystem->has($entity->getFilename())) {
                 $this->filesystem->delete($entity->getFilename());
             }
             $entity->setFilename(null);
